requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
starlette==0.27.0
python-multipart==0.0.6

# LangChain Ecosystem - Locked compatible versions
langchain==0.2.17
langchain-core==0.2.43
langchain-community==0.2.19
langchain-groq==0.1.10
langchain-openai==0.1.25
langchain-text-splitters==0.2.4
langchain-chroma==0.1.2

# LLM APIs
openai==1.54.0
groq==0.11.0

# Embeddings & Transformers
huggingface-hub==0.25.2
transformers==4.46.0
sentence-transformers==3.2.1
tokenizers==0.20.3
tiktoken==0.8.0

# Vector Database - FIXED VERSIONS (Critical: exact versions required)
chromadb==0.4.18
chroma-hnswlib==0.7.3
# DO NOT specify hnswlib directly - let chroma-hnswlib manage it

# Web Scraping
beautifulsoup4==4.12.2
lxml==4.9.3

# HTTP & Networking - CRITICAL UPDATES for compatibility
requests==2.32.3
httpx==0.27.2
urllib3==2.2.3

# Data Validation - CRITICAL UPDATES
pydantic==2.10.3
pydantic-core==2.27.1
pydantic-settings==2.7.0

# Environment & Config
python-dotenv==1.0.1

# UI Framework
streamlit==1.40.2

# File System & Storage - Fix fsspec conflict
fsspec==2024.9.0

# Data Processing - Fix pyarrow conflict
pyarrow==18.1.0

# CLI Tools - Fix click conflict
click==8.1.7

# Image Processing - Fix Pillow conflict
Pillow==10.2.0

# Additional Dependencies
numpy==1.26.4
tenacity==8.5.0
jsonpatch==1.33
jsonpointer==3.0.0
typing-extensions==4.12.2
typing-inspect==0.9.0
annotated-types==0.7.0

# Async & Networking
anyio==3.7.1
h11==0.14.0
sniffio==1.3.1

# Caching & Utilities
cachetools==5.5.2
watchdog==4.0.2

# Monitoring
opentelemetry-api==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0